Refactor main window components for improved structure and functionality
- Removed the ItemTile, PanelHeader, and Panel classes from _main_window.py to streamline the codebase.
- Introduced AppConfig import to enhance configuration management.
- Updated the layout and organization of the main window components for better maintainability.
